Switch rtm_id to guint for consistency
Shouldn't hurt, since overflowing the 32-bit id counter with outstanding
requests is unlikely.  Investigating #36.

+++ slack-rtm.c
@@ -103,9 +103,9 @@ static void rtm_cb(PurpleWebsocket *ws, gpointer data, PurpleWebsocketOp op, con
 	const char *type = json_get_prop_strptr(json, "type");
 
 	if (reply_to) {
-		SlackRTMCall *call = g_hash_table_lookup(sa->rtm_call, GUINT_TO_POINTER((gulong) reply_to->u.integer));
+		SlackRTMCall *call = g_hash_table_lookup(sa->rtm_call, GUINT_TO_POINTER((guint) reply_to->u.integer));
 		if (call) {
-			g_hash_table_steal(sa->rtm_call, GUINT_TO_POINTER((gulong) reply_to->u.integer));
+			g_hash_table_steal(sa->rtm_call, GUINT_TO_POINTER((guint) reply_to->u.integer));
 			if (!json_get_prop_boolean(json, "ok", FALSE)) {
 				json_value *err = json_get_prop(json, "error");
 				if (err->type == json_object)
@@ -176,10 +176,10 @@ void slack_rtm_cancel(SlackRTMCall *call) {
 }
 
 void slack_rtm_send(SlackAccount *sa, SlackRTMCallback *callback, gpointer user_data, const char *type, ...) {
-	gulong id = ++sa->rtm_id;
+	guint id = ++sa->rtm_id;
 
 	GString *json = g_string_new(NULL);
-	g_string_printf(json, "{\"id\":%lu,\"type\":\"%s\"", id, type);
+	g_string_printf(json, "{\"id\":%u,\"type\":\"%s\"", id, type);
 	va_list qargs;
 	va_start(qargs, type);
 	const char *key;
